adjust purchase, stockInOut page code etc

+++ microservice-provider-erp/src/main/java/com/hzg/erp/ErpService.java
@@ -1,21 +1,16 @@
 package com.hzg.erp;
 
 import com.google.common.reflect.TypeToken;
-import com.hzg.order.Order;
-import com.hzg.order.OrderDetail;
-import com.hzg.order.OrderDetailProduct;
 import com.hzg.pay.Account;
 import com.hzg.pay.Pay;
 import com.hzg.pay.Refund;
 import com.hzg.sys.*;
 import com.hzg.tools.*;
-import org.apache.commons.collections.map.HashedMap;
 import org.apache.log4j.Logger;
 import org.hibernate.SessionFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.lang.reflect.Field;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.sql.Timestamp;
@@ -108,13 +103,13 @@ public class ErpService {
     }
 
     public String queryProductOnSalePreFlowAuditNo(Product product){
-        PurchaseDetail purchaseDetail = new PurchaseDetail();
-        purchaseDetail.setProduct(product);
-        Purchase purchase = ((PurchaseDetail)erpDao.query(purchaseDetail).get(0)).getPurchase();
+        PurchaseDetailProduct detailProduct = new PurchaseDetailProduct();
+        detailProduct.setProduct(product);
+        PurchaseDetail detail = ((PurchaseDetailProduct)erpDao.query(detailProduct).get(0)).getPurchaseDetail();
 
         Audit audit = new Audit();
         audit.setEntity(Purchase.class.getSimpleName().toLowerCase());
-        audit.setEntityId(purchase.getId());
+        audit.setEntityId(detail.getPurchase().getId());
 
         List<Audit> dbAudits = writer.gson.fromJson(
                 sysClient.query(Audit.class.getSimpleName().toLowerCase(), writer.gson.toJson(audit)),
@@ -180,7 +175,7 @@ public class ErpService {
                 Product product = detail.getProduct();
 
                 detail.setProduct(product);
-                detail.setNo(product.getNo());
+                detail.setProductNo(product.getNo());
                 detail.setProductName(product.getName());
                 detail.setAmount(product.getUnitPrice() * detail.getQuantity());
                 detail.setPrice(product.getUnitPrice());
@@ -204,10 +199,10 @@ public class ErpService {
                     productQuantity = 1;
                 }
 
-                for (int i = 0; i < productQuantity; i++) {
-                    ProductDescribe describe = product.getDescribe();
-                    result += erpDao.insert(describe);
+                ProductDescribe describe = product.getDescribe();
+                result += erpDao.save(describe);
 
+                for (int i = 0; i < productQuantity; i++) {
                     product.setDescribe(describe);
                     result += erpDao.insert(product);
 
@@ -239,6 +234,83 @@ public class ErpService {
         return result.equals(CommonConstant.fail) ? result : result.substring(CommonConstant.fail.length());
     }
 
+    public String saveProductsCheckDetail(ProductCheck productCheck) {
+        String result = CommonConstant.fail;
+        // 新建一个商品盘点单并设置该商品盘点单的id为productCheck的id
+        ProductCheck productCheck1 = new ProductCheck();
+        productCheck1.setId(productCheck.getId());
+        if (productCheck.getDetails() != null && !productCheck.getDetails().isEmpty()) {
+            for (ProductCheckDetail detail : productCheck.getDetails()) {
+                if (detail.getId() != null){
+                    result += erpDao.updateById(detail.getId(),detail);
+                    continue;
+                }
+                // 不存在该条形码的商品
+                if (erpDao.queryById(detail.getProduct().getId(),Product.class) == null){
+                    detail.setProductCheck(productCheck1);
+                    detail.setUnit("件");
+                    result += erpDao.save(detail);
+                    continue;
+                }
+                // 设置盘点详细条目的盘点单据编号
+                detail.setProductCheck(productCheck1);
+                ProductCheckDetail detail2 = new ProductCheckDetail();
+                detail2.setProductCheck(productCheck1);
+                detail2.setProduct(detail.getProduct());
+                // 该条形码所对应的详细条目已存在，说明是商品扫重复了
+                if(erpDao.query(detail2) != null && !erpDao.query(detail2).isEmpty()){
+                    continue;
+                }
+                Product product = detail.getProduct();
+                product = (Product) (erpDao.query(product).get(0));
+                ProductCheckDetail detail1 = new ProductCheckDetail();
+                detail1.setItemNo(product.getNo());
+                detail1.setProductCheck(productCheck1);
+                // 该存货编码所对应的盘点详细条目已存在
+                if (erpDao.query(detail1) != null && !erpDao.query(detail1).isEmpty()){
+                    detail1 = (ProductCheckDetail)(erpDao.query(detail1).get(0));
+                    // 更新已存在条目的盘点数量及盘点金额
+                    detail1.setCheckQuantity(detail1.getCheckQuantity()+detail.getCheckQuantity());
+                    detail1.setCheckAmount((detail1.getCheckQuantity()+detail.getCheckQuantity())*product.getUnitPrice());
+                    result += erpDao.updateById(detail1.getId(),detail1);
+
+                    // 插入一条盘点数量和盘点金额为0的数据，目的是为了扫描商品时判断是否扫重
+                    detail.setItemNo(detail1.getItemNo());
+                    detail.setItemName(detail1.getItemName());
+                    detail.setCheckQuantity(0.0f);
+                    detail.setCheckAmount(product.getUnitPrice()*0.0f);
+                    detail.setPaperQuantity(detail1.getPaperQuantity());
+                    detail.setPaperAmount(detail1.getPaperAmount());
+                    detail.setUnit(detail1.getUnit());
+                    detail.setUnitPrice(detail1.getUnitPrice());
+                    result += erpDao.save(detail);
+
+                }else{
+                    Stock stock = new Stock();
+                    stock.setProductNo(product.getNo());
+                    stock.setWarehouse(productCheck.getWarehouse());
+                    // 该仓库中存在该商品编码的商品
+                    if (erpDao.query(stock) != null && !erpDao.query(stock).isEmpty()){
+                        stock = (Stock)(erpDao.query(stock).get(0));
+                        detail.setPaperQuantity(stock.getQuantity());
+                        detail.setPaperAmount(stock.getQuantity()*product.getUnitPrice());
+                    }
+                    PurchaseDetail purchaseDetail = new PurchaseDetail();
+                    purchaseDetail.setProductNo(product.getNo());
+                    purchaseDetail = (PurchaseDetail)(erpDao.query(purchaseDetail).get(0));
+                    detail.setUnit(purchaseDetail.getUnit());
+                    detail.setUnitPrice(product.getUnitPrice());
+                    detail.setItemNo(product.getNo());
+                    detail.setItemName(product.getName());
+                    detail.setCheckAmount(detail.getCheckQuantity()*product.getUnitPrice());
+                    result += erpDao.save(detail);
+                }
+            }
+        }
+
+        return result.equals(CommonConstant.fail) ? result : result.substring(CommonConstant.fail.length());
+    }
+
     public String deletePurchaseProducts(Purchase purchase) {
         String result = CommonConstant.fail;
 
@@ -246,8 +318,9 @@ public class ErpService {
             for (PurchaseDetail detail : purchase.getDetails()) {
                 PurchaseDetail dbDetail = (PurchaseDetail) erpDao.queryById(detail.getId(), detail.getClass());
 
+                Product product = null;
                 for (PurchaseDetailProduct detailProduct : dbDetail.getPurchaseDetailProducts()) {
-                    Product product = (Product) erpDao.queryById(detailProduct.getProduct().getId(), detailProduct.getProduct().getClass());
+                    product = (Product) erpDao.queryById(detailProduct.getProduct().getId(), detailProduct.getProduct().getClass());
 
                     if (product.getProperties() != null && !product.getProperties().isEmpty()) {
                         for (ProductOwnProperty ownProperty : product.getProperties()) {
@@ -255,8 +328,11 @@ public class ErpService {
                         }
                     }
 
-                    result += erpDao.delete(product.getDescribe());
                     result += erpDao.delete(product);
+                }
+
+                if (product != null) {
+                    result += erpDao.delete(product.getDescribe());
                 }
 
                 result += erpDao.delete(detail);
@@ -513,15 +589,6 @@ public class ErpService {
         }
 
         return result.equals(CommonConstant.fail) ? result : result.substring(CommonConstant.fail.length());
-    }
-
-    public PurchaseDetail getPurchaseDetail(Integer productId) {
-        PurchaseDetail tempPurchaseDetail = new PurchaseDetail();
-        Product tempProduct = new Product();
-        tempProduct.setId(productId);
-        tempPurchaseDetail.setProduct(tempProduct);
-
-        return (PurchaseDetail) erpDao.query(tempPurchaseDetail).get(0);
     }
 
     private String setStockQuantity(Stock stock, Float quantity, String operator) {
@@ -879,6 +946,26 @@ public class ErpService {
         } else if (entity.equalsIgnoreCase(ProductDescribe.class.getSimpleName())) {
             return erpDao.queryBySql(getProductDescribeSql(json, position, rowNum), ProductDescribe.class);
 
+        } else if (entity.equalsIgnoreCase(ProductCheck.class.getSimpleName())) {
+            Class[] clazzs = {ProductCheck.class, Warehouse.class, Dept.class, User.class, Company.class};
+            Map<String, List<Object>> results = erpDao.queryBySql(getProductCheckComplexSql(json, position, rowNum), clazzs);
+
+            List<Object> productChecks = results.get(ProductCheck.class.getName());
+            List<Object> warehouses = results.get(Warehouse.class.getName());
+            List<Object> depts = results.get(Dept.class.getName());
+            List<Object> users = results.get(User.class.getName());
+            List<Object> companies = results.get(Company.class.getName());
+
+            int i = 0;
+            for (Object productCheck : productChecks) {
+                ((ProductCheck)productCheck).setWarehouse((Warehouse) warehouses.get(i));
+                ((ProductCheck)productCheck).setDept((Dept) depts.get(i));
+                ((ProductCheck)productCheck).setChartMaker((User) users.get(i));
+                ((ProductCheck)productCheck).setCompany((Company) companies.get(i));
+
+                i++;
+            }
+            return productChecks;
         }
 
         return new ArrayList();
@@ -896,6 +983,8 @@ public class ErpService {
         } else if (entity.equalsIgnoreCase(ProductDescribe.class.getSimpleName())) {
             sql = getProductDescribeSql(json, 0, -1);
 
+        }else if (entity.equalsIgnoreCase(ProductCheck.class.getSimpleName())) {
+            sql = getProductCheckComplexSql(json, 0, -1);
         }
 
         sql = "select count(t.id) from " + sql.split(" from ")[1];
@@ -1041,6 +1130,61 @@ public class ErpService {
                 }
             }
 
+
+            if (whereSql.indexOf(" and") == 0) {
+                whereSql = whereSql.substring(" and".length());
+            }
+
+            sql = "select " + selectSql + " from " + fromSql + " where " + whereSql + " order by " + sortNumSql;
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        return sql;
+    }
+
+    private String getProductCheckComplexSql(String json, int position, int rowNum) {
+        String sql = "";
+
+        try {
+            Map<String, Object> queryParameters = writer.gson.fromJson(json, new com.google.gson.reflect.TypeToken<Map<String, Object>>() {}.getType());
+
+            String productCheckSql = objectToSql.generateComplexSqlByAnnotation(ProductCheck.class,
+                    writer.gson.fromJson(writer.gson.toJson(queryParameters.get("productCheck")),
+                            new com.google.gson.reflect.TypeToken<Map<String, String>>() {}.getType()), position, rowNum);
+
+            String selectSql = "", fromSql = "", whereSql = "", sortNumSql = "";
+
+            String[] sqlParts = erpDao.getSqlPart(productCheckSql, ProductCheck.class);
+            selectSql = sqlParts[0];
+            fromSql = sqlParts[1];
+            whereSql = sqlParts[2];
+            sortNumSql = sqlParts[3];
+
+            selectSql += ", " + erpDao.getSelectColumns("t11", Warehouse.class);
+            fromSql += ", " + objectToSql.getTableName(Warehouse.class) + " t11 ";
+            if (!whereSql.trim().equals("")) {
+                whereSql += " and ";
+            }
+            whereSql += " t11." + objectToSql.getColumn(Warehouse.class.getDeclaredField("id")) +
+                    " = t." + objectToSql.getColumn(ProductCheck.class.getDeclaredField("warehouse"));
+
+            selectSql += ", " + erpDao.getSelectColumns("t12", Dept.class);
+            fromSql += ", " + objectToSql.getTableName(Dept.class) + " t12 ";
+            whereSql += " and t12." + objectToSql.getColumn(Dept.class.getDeclaredField("id")) +
+                    " = t." + objectToSql.getColumn(ProductCheck.class.getDeclaredField("dept"));
+
+            selectSql += ", " + erpDao.getSelectColumns("t13", User.class);
+            fromSql += ", " + objectToSql.getTableName(User.class) + " t13 ";
+            whereSql += " and t13." + objectToSql.getColumn(User.class.getDeclaredField("id")) +
+                    " = t." + objectToSql.getColumn(ProductCheck.class.getDeclaredField("chartMaker"))+
+                    " and t13." + objectToSql.getColumn(User.class.getDeclaredField("name")) +
+            " like '%" + ((Map)(queryParameters.get("chartMaker"))).get("name")+"%'";
+
+            selectSql += ", " + erpDao.getSelectColumns("t14", Company.class);
+            fromSql += ", " + objectToSql.getTableName(Company.class) + " t14 ";
+            whereSql += " and t14." + objectToSql.getColumn(Company.class.getDeclaredField("id")) +
+                    " = t." + objectToSql.getColumn(ProductCheck.class.getDeclaredField("company"));
 
             if (whereSql.indexOf(" and") == 0) {
                 whereSql = whereSql.substring(" and".length());
